Add sphinxEnv to set environment variables for sphinx-build process.
Usage would be to pass values from build to sphinx.  The -D option to
sphinx doesn't apply within conf.py, so the values provided in this
way cannot be used there.

Alternatives to using environment variables are to duplicate values in
the build and in conf.py, to preprocess conf.py, preprocess sphinx sources,
postprocess generated files, generate files for conf.py to read from, ...

With this approach:

build.sbt:

 sphinxEnv in Sphinx ++= Map(
   "key" -> "value"
 )

conf.py

 key = os.environ["key"]